memberId 검증 로직 구현

+++ practice/src/main/java/com/solo/practice/member/service/MemberService.java
@@ -1,5 +1,6 @@
 package com.solo.practice.member.service;
 
+import com.solo.practice.auth.userDetailsService.CustomUserDetails;
 import com.solo.practice.auth.utils.CustomAuthorityUtils;
 import com.solo.practice.exception.BusinessLogicException;
 import com.solo.practice.exception.ExceptionCode;
@@ -55,6 +56,13 @@ public class MemberService {
         }
     }
 
+    public void checkMemberId(long memberId, CustomUserDetails customUserDetails){
+
+        if(memberId != customUserDetails.getMemberId()){
+            throw new BusinessLogicException(ExceptionCode.MEMBER_NOT_MATCHED);
+        }
+    }
+
     public Member createMember(Member member){
 
         verifyExistsEmail(member.getEmail());
@@ -69,9 +77,9 @@ public class MemberService {
         return memberRepository.save(member);
     }
 
-    public Member updateMember(Member member){
+    public Member updateMember(Member member, CustomUserDetails customUserDetails){
 
-        Member findMember = findVerifiedMember(member.getMemberId());
+        Member findMember = findVerifiedMember(customUserDetails.getMemberId());
 
         Optional.ofNullable(member.getNickname())
                 .ifPresent(nickname -> {
@@ -84,13 +92,19 @@ public class MemberService {
         return memberRepository.save(findMember);
     }
 
-    public Member findMember(long memberId){
-        return findVerifiedMember(memberId);
+    public Member findMember(CustomUserDetails customUserDetails){
+
+        return findVerifiedMember(customUserDetails.getMemberId());
     }
 
-    public void deleteMember(long memberId){
+    public void deleteMember(String password, CustomUserDetails customUserDetails){
 
-        Member findMember = findVerifiedMember(memberId);
+        Member findMember = findVerifiedMember(customUserDetails.getMemberId());
+
+        if(!passwordEncoder.matches(password, findMember.getPassword())){
+            throw new BusinessLogicException(ExceptionCode.MEMBER_PASSWORD_NOT_MATCHED);
+        }
+
         memberRepository.delete(findMember);
     }
 }
